pyproject.toml
[project]
name = "ecommerce-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Melissa1221",email = "melissa.iman.n@uni.pe"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain-core (>=0.3.31,<0.4.0)",
    "langgraph (>=0.2.67,<0.3.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
